move inito detail
